+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add new api client for Granite api calls
+- Add new resource to manage SSL via Granite
+
 ### Changed
 - Upgrade swagger_aem to 3.3.0